move api-client boto3 functions to shared library
